shuffled routes and rewritten somewhat needs testing
