+++ src/api/routers/feed.py
@@ -1,0 +1,67 @@
+from fastapi import APIRouter, Depends
+
+from typing import List
+
+from db.category import Category
+from db.feed import Feed
+from db.user import User
+from route_models.item import ItemResponse
+from route_models.acknowledge import AcknowledgeResponse
+from routers.auth import authenticate
+
+feed_router = APIRouter()
+
+
+@feed_router.post(
+    "/create",
+    summary="Create a feed (within a category)",
+    response_model=AcknowledgeResponse,
+)
+def create_feed(
+    category_name_hash: str,
+    feed_name_hash: str,
+    feed_url: str,
+    user: User = Depends(authenticate),
+) -> AcknowledgeResponse:
+    cat = Category.read(user_hash=user.name_hash, name_hash=category_name_hash)
+    feed = Feed(
+        user_hash=user.name_hash,
+        category_hash=category_name_hash,
+        name=feed_name_hash,
+        url=feed_url,
+    )
+    cat.add_feed(feed)
+    return AcknowledgeResponse()
+
+
+@feed_router.get(
+    "/items", summary="Get all items in a feed", response_model=List[ItemResponse]
+)
+def get_items(
+    category_name_hash: str, feed_name_hash: str, user: User = Depends(authenticate)
+) -> List[ItemResponse]:
+    feed = Feed.read(
+        user_hash=user.name_hash,
+        category_hash=category_name_hash,
+        name_hash=feed_name_hash,
+    )
+    items = feed.get_all_items()
+    return [ItemResponse.from_db_model(i) for i in items]
+
+
+@feed_router.delete(
+    "/delete",
+    summary="Delete a feed",
+    response_model=AcknowledgeResponse,
+)
+def delete_feed(
+    category_name_hash: str, feed_name_hash: str, user: User = Depends(authenticate)
+) -> AcknowledgeResponse:
+    cat = Category.read(user_hash=user.name_hash, name_hash=category_name_hash)
+    feed = Feed.read(
+        user_hash=user.name_hash,
+        category_hash=category_name_hash,
+        name_hash=feed_name_hash,
+    )
+    cat.delete_feed(feed)
+    return AcknowledgeResponse
