+++ src/api/routers/item.py
@@ -7,3 +7,5 @@ item_router = APIRouter()
 # TODO rate an item
 
 # TODO get a preview for an item  (and save it to the DB)
+
+# TODO get item's feed that produced it
